services/product-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
PyMySQL==1.1.1
PyJWT==2.9.0
pydantic-settings==2.5.2
